+++ src/utils/index.ts
@@ -10,3 +10,4 @@ export * from "./model-capabilities";
 export * from "./errors";
 export * from "./function-calling";
 export * from "./utf8-decoder";
+export * from "./message-processing";